Fix shifts (#155)
* fixed minor typos and issues. In particular, made complex exponential take arguments instead of exponents.

* Cleaned multi-resolution operator multiplication to remove weird transpose from aliged case

* fixed minor typos and issues. In particular, made complex exponential take arguments instead of exponents.

* Cleaned multi-resolution operator multiplication to remove weird transpose from aliged case

* Fixed shift for identity case by inverting the sign of shift applied to model indel of inverting axis

* fixed shifts for odd fast shapes when image shape is even

* clean up

* added  argument to match and removed useless  argument

* fixed an issue from bad merging. the complex exponential form of shifts was not updated in Components

* Modification as per PR review from Peter

+++ scarlet/fft.py
@@ -88,7 +88,7 @@ def _get_fft_shape(img1, img2, padding=3, axes=None, max=False):
         else:
             shape = shape1 + shape2
     else:
-        shape = np.zeros(len(axes))
+        shape = np.zeros(len(axes), dtype='int')
         try:
             len(axes)
         except TypeError:
